fizz_buzz_with_a_pink_flamingo/parser.py
from math import sqrt

class Parser:
    def __init__(self, start = 0, end = 100):
        self.range = range(start, end + 1)

    def is_multiple_of(self, n, x):
        return x % n == 0

    def is_fibonacci(self, n):
        lower_condition = self.is_multiple_of(1, sqrt(5*n**2 - 4))
        upper_condition = self.is_multiple_of(1, sqrt(5*n**2 + 4))
        return lower_condition or upper_condition

    def generate_output(self):
        print("This will ultimately return the parsed FizzBuzz With a Pink Flamingo output.")